Merge in 2.0.3-2.2
This was an NMU by Barry deFreese <bdefreese@debian.org> to fix the -lm
linking issue.  He put this into LDADD rather than LDFLAGS, which seems
more correct to me.

On the other hand, I don't see why it would be necessary to patch
Makefile.in in addition to Makefile.am, so I dropped that part of the
patch.

Git-Dch: Ignore