Get rid of rarian-compat B-D
Here is a minimalistic patch that fixes this bug.

The GNOME people have ditched several documentation systems in the
past years; I'm not really sure what is the current one.  Probably
yelp-tools or yelp-something.

Switching to another system is likely to require changes in the manual
itself as well as the installation location.  The approach I suggest,
while suboptimal, will work as long as Yelp is able to display URIs
with the "ghelp:" prefix.

Closes: 885638
Signed-off-by: Richard Laager <rlaager@wiktel.com>