Port to GTK+ 3 and GIO
Attached is a patch which should fix this bug.  As I've never used
this program before and I'm completely unfamiliar with US Savings
Bonds, please check that I haven't broken the core application logic
somehow.

The upstream coding style is not consistent, so I tried to stick to
the predominant coding style that is used for most of the code (rather
than using the specific variation at the places where modifications
are made).  If this is unacceptable, just let me know and I'll
reformat the patch.  Needless to say, I'll also do my best to fix any
other problems/regressions that might occur.

--
Richard Laager made the following changes:
  - Dropped spurious diff header changes to other files
  - Folded in the remainder of debian/patches/libgnomeprint

Closes: 868413
Signed-off-by: Richard Laager <rlaager@wiktel.com>